fix a segfault in the proxy on OS/2
Submitted by:	Brian Havard


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@81830 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/proxy/proxy_cache.c
@@ -846,7 +846,8 @@ int ap_proxy_cache_update(struct cache_req *c, array_header *resp_hdrs,
 	    c->fp = NULL;
 	}
 /* delete the previously cached file */
-	unlink(c->filename);
+        if (c->filename)
+            unlink(c->filename);
 	return DECLINED;	/* send data to client but not cache */
     }
